refactor wallet & broker
transaction and wallet services will be separated for scalability

+++ broker-service/internal/models/response.go
@@ -1,6 +1,6 @@
 package models
 
 type ViewBalanceResponse struct {
-	Name    string `json:"name"`
-	Balance int64  `json:"balance"`
+	Name    string  `json:"name"`
+	Balance float64 `json:"balance"`
 }
